+++ com.unity.formats.alembic/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.4.0-pre.1] - 2023-11-07
+### Added
+- New "Hair" tab in the Alembic Importer window to generate a Hair asset from an Alembic file containing curves (subject to Hair package installation).
 
 ## [2.3.4] - 2023-10-30
 ### Fixed